reorg/rename common artifacts
