+++ bentoml/artifact/README.md
@@ -2,11 +2,11 @@
 
 The `bentoml.artifact` module is now deprecated.
 
-Use`bentoml.frameworks.*` and `bentoml.service.*` instead.
+Use`bentoml.frameworks.*` and `bentoml.service.artifact` instead.
 
 e.g.:
 ```
 from bentoml.frameworks.sklearn import SklearnModelArtifact
-from bentoml.service.common_artifacts import PickleArtifact
 from bentoml.service.artifacts import BentoServiceArtifact
+from bentoml.service.artifacts.common import PickleArtifact
 ```
